pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiagentlab-ild"
version = "0.1.0"
description = "Phase 1 skeleton for ILD MDT input-structuring layer"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Team" }]
dependencies = [
	"pydantic>=2,<3",
	"openai",
	"pyyaml",
	"python-dotenv",
]

